+++ packData/packData.mjs
@@ -25,7 +25,8 @@ let packs = [
     'cpr-summon-features-2024',
     'cpr-feats-2024',
     //'cpr-feat-features-2024',
-    'cpr-embedded-macro-sample-items'
+    'cpr-embedded-macro-sample-items',
+    'cpr-actions-2024'
 ];
 for (let i of packs) {
     await compilePack('./packData/' + i, './packs/' + i, {'log': true});