+++ packData/unpackData.mjs
@@ -23,7 +23,8 @@ let itemPacks = [
     'cpr-summon-features-2024',
     'cpr-feats-2024',
     'cpr-feat-features-2024',
-    'cpr-embedded-macro-sample-items'
+    'cpr-embedded-macro-sample-items',
+    'cpr-actions-2024'
 ];
 let actorPacks = [
     'cpr-summons',