+++ scripts/macros/2014/actions/dash.js
@@ -1,4 +1,8 @@
+import {dash as modernDash} from '../../2024/actions/dash.js';
 export let dash = {
     name: 'Dash',
-    version: '1.1.0'
+    version: '1.3.34',
+    midi: modernDash.midi,
+    config: modernDash.config,
+    hasAnimation: true
 };